chore: refactor codes chore: use babel instead of tsc and ts-node
