+++ .mocharc.json
@@ -1,5 +1,5 @@
 {
   "extension": ["ts"],
   "spec": "tests/**/*.test.ts",
-  "require": "ts-node/register"
+  "require": "./tests/babel-register.js"
 }
