+++ package.json
@@ -11,13 +11,15 @@
     "test:load": "artillery run ./artillery/test.yaml",
     "prestart": "npm run build",
     "start": "node ./dist/index.js",
-    "start:dev": "ts-node ./src/index.ts",
+    "start:dev": "babel-node --extensions \".ts,.tsx\" ./src/index.ts",
     "prebuild": "rm -rf dist",
-    "build": "tsc --declaration -p tsconfig-build.json",
-    "prettier": "prettier --check 'src/**/*.*'",
-    "prettier:write": "prettier --write 'src/**/*.*'",
-    "lint": "tslint 'src/**/*.{j,t}s{,x}' && eslint 'src/**/*.{j,t}s{,x}'",
-    "lint:fix": "tslint --fix 'src/**/*.{j,t}s{,x}' && eslint --fix 'src/**/*.{j,t}s{,x}'"
+    "build": "npm run build:types && npm run build:js",
+    "build:types": "tsc --emitDeclarationOnly -p tsconfig-build.json",
+    "build:js": "babel src --out-dir dist --extensions \".ts,.tsx\" --source-maps true",
+    "prettier": "prettier --check '{src,tests}/**/*.*'",
+    "prettier:write": "prettier --write '{src,tests}/**/*.*'",
+    "lint": "tslint '{src,tests}/**/*.{t}s{,x}' && eslint '{src,tests}/**/*.{j,t}s{,x}'",
+    "lint:fix": "tslint --fix '{src,tests}/**/*.{t}s{,x}' && eslint --fix '{src,tests}/**/*.{j,t}s{,x}'"
   },
   "keywords": [
     "Xendit"
@@ -33,6 +35,12 @@
     "swagger-ui-express": "^4.1.4"
   },
   "devDependencies": {
+    "@babel/cli": "^7.10.3",
+    "@babel/core": "^7.10.3",
+    "@babel/node": "^7.10.3",
+    "@babel/preset-env": "^7.10.3",
+    "@babel/preset-typescript": "^7.10.1",
+    "@babel/register": "^7.10.3",
     "@types/body-parser": "^1.19.0",
     "@types/chai": "^4.2.11",
     "@types/cors": "^2.8.6",
@@ -47,16 +55,16 @@
     "@typescript-eslint/eslint-plugin-tslint": "^3.4.0",
     "@typescript-eslint/parser": "^3.4.0",
     "artillery": "^1.6.0",
+    "babel-plugin-module-resolver": "^4.0.0",
     "chai": "^4.2.0",
     "eslint": "^7.3.1",
     "eslint-config-prettier": "^6.11.0",
     "husky": "^4.2.5",
     "lint-staged": "^10.2.11",
-    "mocha": "^6.2.3",
+    "mocha": "^8.0.1",
     "nyc": "^15.1.0",
     "prettier": "^2.0.5",
     "supertest": "^4.0.2",
-    "ts-node": "^8.10.2",
     "tslint": "^6.1.2",
     "tslint-microsoft-contrib": "^6.2.0",
     "typescript": "^3.9.5"
