+++ src/constants.ts
@@ -1,0 +1,3 @@
+export const VALIDATION_ERROR = "VALIDATION_ERROR";
+export const RIDES_NOT_FOUND_ERROR = "RIDES_NOT_FOUND_ERROR";
+export const SERVER_ERROR = "SERVER_ERROR";
