--- tests/api.test.ts
@@ -1,336 +1,0 @@
-import { expect } from "chai";
-import { Application } from "express";
-import request, { Test } from "supertest";
-import app, { RIDES_NOT_FOUND_ERROR, VALIDATION_ERROR } from "../src/app";
-import getDB, { DB } from "../src/getDB";
-import buildSchemas from "../src/schemas";
-import { RideInput } from "../src/types";
-
-function requestToPromise(req: Test): Promise<void> {
-  return new Promise((resolve, reject) =>
-    req.end((err) => {
-      if (err) {
-        return reject(err);
-      }
-      resolve();
-    }),
-  );
-}
-
-describe("API tests", () => {
-  let db: DB, application: Application;
-
-  before(async () => {
-    db = getDB();
-    await db.serialize();
-    await buildSchemas(db);
-    application = app(db);
-  });
-
-  describe("GET /health", () => {
-    it("should return health", (done) => {
-      request(application)
-        .get("/health")
-        .expect("Content-Type", /text/)
-        .expect(200, done);
-    });
-  });
-
-  describe("Rides", () => {
-    const payload: RideInput = {
-      startLat: 10,
-      startLong: 20,
-      endLat: 30,
-      endLong: 40,
-      riderName: "a",
-      driverName: "b",
-      driverVehicle: "c",
-    };
-
-    before(async () => {
-      // prepopolate 3 rides
-      await requestToPromise(request(application).post("/rides").send(payload));
-      await requestToPromise(request(application).post("/rides").send(payload));
-      await requestToPromise(request(application).post("/rides").send(payload));
-    });
-
-    describe("POST /rides", () => {
-      it("should return the ride after POST if success", (done) => {
-        request(application)
-          .post("/rides")
-          .send(payload)
-          .expect("Content-Type", /application\/json/)
-          .expect((res) => {
-            const rideID = 4;
-            expect(res.body.rideID).to.eql(rideID);
-            expect(res.body).to.eql({
-              rideID,
-              ...payload,
-              created: res.body.created,
-            });
-          })
-          .expect(200, done);
-      });
-
-      it("should return error code if startLat not in range", async () => {
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              startLat: -91,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              startLat: 91,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-      });
-
-      it("should return error code if startLong not in range", async () => {
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              startLong: -181,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              startLong: 181,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-      });
-
-      it("should return error code if endLat not in range", async () => {
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              endLat: -91,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              endLat: 91,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-      });
-
-      it("should return error code if endLong not in range", async () => {
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              endLong: -181,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              endLong: 181,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-      });
-
-      it("should return error code if riderName not valid", async () => {
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              riderName: "",
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              riderName: 1,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-      });
-
-      it("should return error code if driverName not valid", async () => {
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              driverName: "",
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              driverName: 1,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-      });
-
-      it("should return error code if driverVehicle not valid", async () => {
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              driverVehicle: "",
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-        await requestToPromise(
-          request(application)
-            .post("/rides")
-            .send({
-              ...payload,
-              driverVehicle: 1,
-            })
-            .expect("Content-Type", /application\/json/)
-            .expect((res) => {
-              expect(res.body["error_code"]).to.eql(VALIDATION_ERROR);
-            })
-            .expect(422),
-        );
-      });
-    });
-
-    describe("GET /rides", () => {
-      it("should get all rides without offset / limit", (done) => {
-        request(application)
-          .get("/rides")
-          .expect("Content-Type", /application\/json/)
-          .expect((res) => {
-            expect(res.body.length).to.greaterThan(3);
-          })
-          .expect(200, done);
-      });
-
-      it("should get rides with offset / limit", (done) => {
-        request(application)
-          .get("/rides?offset=1&limit=2")
-          .expect("Content-Type", /application\/json/)
-          .expect(({ body }) => {
-            expect(body.length).to.eql(2);
-            // should not able to find first ride
-            const found = body.find((r) => r.rideID === 1);
-            expect(found).to.eql(undefined, "should not contain first ride");
-          })
-          .expect(200, done);
-      });
-
-      it("should return error if cannot find any", (done) => {
-        request(application)
-          .get("/rides?offset=100&limit=2")
-          .expect("Content-Type", /application\/json/)
-          .expect(({ body }) => {
-            expect(body["error_code"]).to.eql(RIDES_NOT_FOUND_ERROR);
-          })
-          .expect(404, done);
-      });
-    });
-
-    describe("GET /rides/{id}", () => {
-      it("should return ride if found the id", (done) => {
-        request(application)
-          .get("/rides/2")
-          .expect("Content-Type", /application\/json/)
-          .expect(({ body }) => {
-            expect(body.rideID).to.eql(2);
-          })
-          .expect(200, done);
-      });
-
-      it("should return error if cannot find any", (done) => {
-        request(application)
-          .get("/rides/1000")
-          .expect("Content-Type", /application\/json/)
-          .expect(({ body }) => {
-            expect(body["error_code"]).to.eql(RIDES_NOT_FOUND_ERROR);
-          })
-          .expect(404, done);
-      });
-    });
-  });
-});
