+++ tests/babel-register.js
@@ -1,0 +1,2 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+require("@babel/register")({ extensions: [".js", ".jsx", ".ts", ".tsx"] });
