+++ tsconfig.json
@@ -1,9 +1,10 @@
 {
   "compilerOptions": {
+    "target": "es6",
     "module": "commonjs",
+    "declaration": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
-    "target": "es6",
     // "noImplicitAny": true,
     "moduleResolution": "node",
     "sourceMap": true,
@@ -11,7 +12,7 @@
     "baseUrl": ".",
     "paths": {
       "*": ["node_modules/*"],
-      "../*": ["src/*", "dist/*"]
+      "~/*": ["src/*"]
     },
     "typeRoots": ["node_modules/@types", "src/types"],
     "strictNullChecks": true,
